requirements.txt
faiss-cpu
sentence-transformers
PyPDF2
python-docx
beautifulsoup4
openai
tqdm